requirements.txt
# Core dependencies
tinker-toolkit>=0.1.0
python-dotenv>=1.0.0
cohere>=5.0.0

# ML/Data processing
torch>=2.0.0
numpy>=1.24.0
transformers>=4.35.0
huggingface-hub>=0.19.0

# Utilities
tqdm>=4.65.0

